+++ src/memoEdit.ts
@@ -8,7 +8,7 @@ import * as nls from 'vscode-nls';
 import * as os from 'os';
 import { items, memoConfigure } from './memoConfigure';
 
-const localize = nls.config(process.env.VSCODE_NLS_CONFIG)();
+const localize = nls.config({ messageFormat: nls.MessageFormat.file })();
 
 export class memoEdit extends memoConfigure  {
     public memoListChannel: vscode.OutputChannel;